useRepositoryProjectMapping -> useCodeMappings
